refactor: 외부 라이브러리 레이어 변경
