+++ src/main/java/wooteco/subway/dto/path/PathResponse.java
@@ -2,7 +2,7 @@ package wooteco.subway.dto.path;
 
 import wooteco.subway.domain.Station;
 import wooteco.subway.domain.fare.Fare;
-import wooteco.subway.domain.path.Path;
+import wooteco.subway.domain.Path;
 import wooteco.subway.dto.station.StationResponse;
 
 import java.util.List;
