+++ src/main/java/wooteco/subway/infra/path/CustomEdge.java
@@ -1,4 +1,4 @@
-package wooteco.subway.domain;
+package wooteco.subway.infra.path;
 
 import org.jgrapht.graph.DefaultWeightedEdge;
